+++ src/main/java/com/woosan/hr_system/survey/controller/SurveyApiController.java
@@ -1,15 +1,13 @@
 package com.woosan.hr_system.survey.controller;
 
+import com.woosan.hr_system.aspect.RequireManagerPermission;
 import com.woosan.hr_system.survey.model.Response;
 import com.woosan.hr_system.survey.model.Survey;
 import com.woosan.hr_system.survey.service.SurveyService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -20,6 +18,7 @@ public class SurveyApiController {
     @Autowired
     private SurveyService surveyService;
 
+    @RequireManagerPermission
     @PostMapping // 새로운 설문 제출
     public ResponseEntity<String> submitSurvey(@RequestBody Survey survey) {
         return ResponseEntity.ok(surveyService.submitSurvey(survey));
@@ -27,7 +26,12 @@ public class SurveyApiController {
 
     @PostMapping("/response") // 설문 응답 제출
     public ResponseEntity<String> submitResponse(@RequestBody List<Response> responses) {
-        log.debug("{}", responses);
         return ResponseEntity.ok(surveyService.submitResponse(responses));
+    }
+
+    @RequireManagerPermission
+    @DeleteMapping("/{id}") // 설문 삭제
+    public ResponseEntity<String> deleteSurvey(@PathVariable("id") int id) {
+        return ResponseEntity.ok(surveyService.deleteSurvey(id));
     }
 }
