+++ src/main/java/com/woosan/hr_system/survey/dao/SurveyDAO.java
@@ -93,4 +93,9 @@ public class SurveyDAO {
         map.put("employeeId", employeeId);
         return sqlSession.selectOne("survey.selectParticipationInfo", map);
     }
+
+    // 설문 삭제
+    public void deleteSurvey(int id) {
+        sqlSession.delete("survey.deleteSurvey", id);
+    }
 }
